requirements.txt
alembic==1.13.2
Flask==3.1.0
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
python_dateutil==2.9.0.post0
SQLAlchemy==2.0.31
Werkzeug==3.1.3
